+++ doc/release-notes.md
@@ -23,6 +23,9 @@ Low-level RPC Changes
 - The RPC gettransaction, listtransactions and listsinceblock responses now also
   includes the height of the block that contains the wallet transaction, if any.
 
+- A new descriptor type `sortedmulti(...)` has been added to support multisig scripts
+  where the public keys are sorted lexicographically in the resulting script.
+
 Deprecated or removed RPCs
 --------------------------
 - The `getaddressinfo` RPC `labels` field now returns an array of label name